% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2tidymodels.R
\name{fastkrr_fit_wrapper}
\alias{fastkrr_fit_wrapper}
\title{Internal bridge for parsnip fit interface}
\usage{
fastkrr_fit_wrapper(
  x,
  y,
  kernel = "gaussian",
  opt = "exact",
  m = NULL,
  eps = 1e-06,
  rho = 1,
  lambda = NULL,
  n_threads = 4,
  verbose = FALSE,
  na.rm = FALSE
)
}
\description{
Internal bridge for parsnip fit interface
}
\keyword{internal}
