% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param}
\alias{param}
\alias{param.default}
\title{Extract/print hyperparameters of fitted models}
\usage{
param(x, ...)

\method{param}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list of hyperparameters (invisibly); may print side effects.
}
\description{
\code{"param()"} is a generic S3 function that displays (and invisibly returns)
model hyperparameters. Methods are provided for \code{"krr"} objects.
}
