% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2tidymodels.R
\name{tunable.krr_reg}
\alias{tunable.krr_reg}
\title{Expose tunable parameters for `krr_reg`}
\usage{
\method{tunable}{krr_reg}(x, ...)
}
\arguments{
\item{x}{A `krr_reg` model specification.}

\item{...}{Not used; included for S3 method compatibility.}
}
\value{
A tibble (one row per tunable parameter)
   with columns 'name', 'call_info', 'source',
  'component', and 'component_id'.
}
\description{
Supplies a tibble of tunable arguments for `krr_reg()`.
}
