% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2tidymodels.R
\name{tunable.krr_reg}
\alias{tunable.krr_reg}
\title{Expose tunable parameters for \code{"krr_reg"}}
\usage{
\method{tunable}{krr_reg}(x, ...)
}
\arguments{
\item{x}{A \code{"krr_reg"} model specification.}

\item{...}{Not used; included for S3 method compatibility.}
}
\value{
A tibble (one row per tunable parameter)
   with columns \code{"name"}, \code{"call_info"}, \code{"source"},
  \code{"component"}, and \code{"component_id"}.
}
\description{
Supplies a tibble of tunable arguments for \code{"krr_reg()"}.
}
